Refactor: hooks to request data with pagination
